split diskstorage tests
