+++ .circleci/config.yml
@@ -33,7 +33,8 @@ jobs:
                   paths:
                       - ~/.m2
                   key: v1-dependencies-{{ checksum "pom.xml" }}
-    test-cql-diskstorage:
+
+    test-cql-diskstorage-I:
         docker:
             - image: circleci/openjdk:8u222-jdk-stretch
         working_directory: ~/repo
@@ -54,7 +55,30 @@ jobs:
                   key: v1-dependencies-{{ checksum "pom.xml" }}
 
             # run tests!
-            - run: mvn -q test -pl janusgraph-cql '-Dtest=org.janusgraph.diskstorage.cql.*Test'
+            - run: mvn -q test -pl janusgraph-cql '-Dtest=CQLDistributedStoreManagerTest,CQLIDAuthorityTest,CQLLockStoreTest,CQLLogTest'
+
+    test-cql-diskstorage-II:
+        docker:
+            - image: circleci/openjdk:8u222-jdk-stretch
+        working_directory: ~/repo
+        environment:
+            MAVEN_OPTS: -Xmx3200m
+        steps:
+            - checkout
+            - restore_cache:
+                  keys:
+                      - v1-dependencies-{{ checksum "pom.xml" }}
+                      # fallback to using the latest cache if no exact match is found
+                      - v1-dependencies-
+
+            - run: mvn install -DskipTests
+            - save_cache:
+                  paths:
+                      - ~/.m2
+                  key: v1-dependencies-{{ checksum "pom.xml" }}
+
+            # run tests!
+            - run: mvn -q test -pl janusgraph-cql '-Dtest=CQLMultiWriteStoreTest,CQLResultSetKeyIteratorTest,CQLSSLStoreTest,CQLStoreTest'
 
     test-cql-graphdb-I:
         docker:
@@ -129,7 +153,8 @@ workflows:
     version: 2
     grakn-janusgraph:
         jobs:
-            - test-cql-diskstorage
+            - test-cql-diskstorage-I
+            - test-cql-diskstorage-II
             - test-cql-graphdb-I
             - test-cql-graphdb-II
 #            - test-cql-blueprints Running these means running > 2400 tests - they fail after 2:50 hours - investigate
