+++ janusgraph-cql/src/test/java/org/janusgraph/diskstorage/cql/utils/CassandraStorageSetup.java
@@ -12,12 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.janusgraph.diskstorage.cql;
+package org.janusgraph.diskstorage.cql.utils;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Preconditions;
 import org.apache.commons.io.FileUtils;
-import org.janusgraph.diskstorage.cql.utils.CassandraDaemonWrapper;
 import org.janusgraph.diskstorage.configuration.ConfigElement;
 import org.janusgraph.diskstorage.configuration.ModifiableConfiguration;
 import org.slf4j.Logger;
