+++ janusgraph-cql/src/test/java/org/janusgraph/graphdb/cql/CQLGraphConcurrentTest.java
@@ -16,7 +16,7 @@ package org.janusgraph.graphdb.cql;
 
 import org.janusgraph.TestCategory;
 import org.janusgraph.diskstorage.configuration.WriteConfiguration;
-import org.janusgraph.diskstorage.cql.CassandraStorageSetup;
+import org.janusgraph.diskstorage.cql.utils.CassandraStorageSetup;
 import org.janusgraph.graphdb.JanusGraphConcurrentTest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
