+++ janusgraph-cql/src/test/java/org/janusgraph/graphdb/cql/CQLOperationCountingTest.java
@@ -15,7 +15,7 @@
 package org.janusgraph.graphdb.cql;
 
 import org.janusgraph.diskstorage.configuration.WriteConfiguration;
-import org.janusgraph.diskstorage.cql.CassandraStorageSetup;
+import org.janusgraph.diskstorage.cql.utils.CassandraStorageSetup;
 import org.janusgraph.graphdb.JanusGraphOperationCountingTest;
 import org.junit.jupiter.api.BeforeAll;
