+++ janusgraph-hadoop-parent/janusgraph-hadoop-core/src/test/java/org/janusgraph/hadoop/CassandraScanJobIT.java
@@ -18,9 +18,8 @@ package org.janusgraph.hadoop;
 import org.janusgraph.core.JanusGraphVertex;
 import org.janusgraph.diskstorage.*;
 import org.janusgraph.diskstorage.configuration.*;
-import org.janusgraph.diskstorage.cql.CQLStoreManager;
 import org.janusgraph.diskstorage.cql.CQLStoreManagerFactory;
-import org.janusgraph.diskstorage.cql.CassandraStorageSetup;
+import org.janusgraph.diskstorage.cql.utils.CassandraStorageSetup;
 import org.janusgraph.diskstorage.keycolumnvalue.KeyColumnValueStore;
 import org.janusgraph.diskstorage.keycolumnvalue.KeyColumnValueStoreManager;
 import org.janusgraph.diskstorage.keycolumnvalue.StoreTransaction;
